refactor: modularize home page sections
