+++ types/home.ts
@@ -1,0 +1,8 @@
+export interface HomeArtistSummary {
+  id: string;
+  name: string;
+  avatarUrl: string | null;
+  bio: string | null;
+  followerCount: number;
+  projectCount: number;
+}
